% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifs.R
\name{build_pair_offsets}
\alias{build_pair_offsets}
\title{Build motif offsets for normalized counts and modeling}
\usage{
build_pair_offsets(Y2, log_vols, log_2m, pseudo)
}
\description{
Build motif offsets for normalized counts and modeling
}
\keyword{internal}
